No more neckless baby horses
Baby horse heads were disconnecting as they turned forward due to the rotation center of the head being offset to account for the neck.  It is possible to remove it again if we want to, but doing so will require different ModelRenderer objects for the adult head and the baby head.

Also removed some comments that no longer apply.