Server code generator added!
When you run the script, it will generate code for the server in a
server/ folder.  Link that in eclipse on the Server project, but be
careful that you don't accidentally start editing the server code
instead of the client code! (speaking from experience here...)

Run the script whenever you're about to test new changes in SMP.

(also, if you're wondering how to test SMP, here's a hint: see
the dropdown arrow next to the Debug button in Eclipse?)