Changed Aversion levels to psuedo-floats
This way aversion levels can be slowly decremented upon exposure. Also
put in the foundation of the 'attack undead when no longer afraid of
them' code. If 8 meters is too short of a distance to be afraid of
threats then we can switch to 1/8 meter increments or switch to using a
short rather than a byte.